Refactor location of ACM
